refactor tests, remove beta messaging and make it GA ready
